fixed typename identifier
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@48825 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsdata/src/aimsgifti/plugin/giftiplugin.cc
@@ -67,7 +67,7 @@ GiftiPlugin::GiftiPlugin() : Plugin()
   FileFormatDictionary<TimeTexture<float> >::registerFormat( "GIFTI",
                                                              df2, ext );
 
-  GiftiTextureFormat<float> *df8 = new GiftiTextureFormat<double>;
+  GiftiTextureFormat<double> *df8 = new GiftiTextureFormat<double>;
   FileFormatDictionary<TimeTexture<double> >::registerFormat( "GIFTI",
                                                              df8, ext );
